Added VisitationRepository class. Resolves #232.
